Fix:[BE] 통계반환 수정, 컨트롤러 토큰 받는 방식 수정

See merge request s11-fintech-finance-sub1/S11P21A304!30

+++ backend/src/main/java/com/f1veguys/sel/domain/spendinghistory/dto/MonthlySpendingDto.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 public class MonthlySpendingDto {
     private int year;
     private int month;
-    private int totalSpending;
-    private int ecoSpending;
+    private long totalSpending;
+    private long ecoSpending;
 
 
 }
